+++ device-bootloaders/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - [SW-14813] Get and display PMIC FW version at SP boot.
+- [SW-14813] Configuring the MPROT ESR for neighboorhoods of all Minion Shires.
 ### Changed
 ### Deprecated
 ### Removed